# contracts/openapi/lead-intelligence-v1.openapi.json
{
  "components": {
    "schemas": {
...
      "DiscoveryEvent": {
        "additionalProperties": false,
        "description": "Tracked discovery request..",
        "properties": {
          "campaign_id": {
            "format": "uuid",
            "title": "Campaign Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "seed_url": {
            "format": "uri",
            "maxLength": 2083,
            "minLength": 1,
            "title": "Seed Url",
            "type": "string"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus",
            "default": "pending"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "workspace_id": {
            "format": "uuid",
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "workspace_id",
          "campaign_id",
          "source_id",
          "seed_url"
        ],
        "title": "DiscoveryEvent",
        "type": "object"
      },
      "DiscoveryEventCreate": {
        "additionalProperties": false,
        "description": "Create request.",
        "properties": {
          "campaign_id": {
            "format": "uuid",
            "title": "Campaign Id",
            "type": "string"
          },
          "seed_url": {
            "format": "uri",
            "maxLength": 2083,
            "minLength": 1,
            "title": "Seed Url",
            "type": "string"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus",
            "default": "pending"
          },
          "workspace_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace Id"
          }
        },
        "required": [
          "campaign_id",
          "source_id",
          "seed_url"
        ],
        "title": "DiscoveryEventCreate",
        "type": "object"
      },
      "DiscoveryEventUpdate": {
        "additionalProperties": false,
        "description": "Partial update request.",
        "properties": {
          "campaign_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Campaign Id"
          },
          "seed_url": {
            "anyOf": [
              {
                "format": "uri",
                "maxLength": 2083,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed Url"
          },
          "source_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Id"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "workspace_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace Id"
          }
        },
        "title": "DiscoveryEventUpdate",
        "type": "object"
      },
      "ExportRecord": {
        "additionalProperties": false,
        "description": "Lead export request and result..",
        "properties": {
          "campaign_id": {
            "format": "uuid",
            "title": "Campaign Id",
            "type": "string"
          },
          "completed_at": {
            "format": "date-time",
            "title": "Completed At",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "destination": {
            "title": "Destination",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "lead_ids": {
            "items": {
              "type": "string"
            },
            "title": "Lead Ids",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/ExportStatus",
            "default": "pending"
          },
          "workspace_id": {
            "format": "uuid",
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "workspace_id",
          "campaign_id",
          "destination",
          "completed_at"
        ],
        "title": "ExportRecord",
        "type": "object"
      },
      "ExportRecordCreate": {
        "additionalProperties": false,
        "description": "Create request.",
        "properties": {
          "campaign_id": {
            "format": "uuid",
            "title": "Campaign Id",
            "type": "string"
          },
          "completed_at": {
            "format": "date-time",
            "title": "Completed At",
            "type": "string"
          },
          "destination": {
            "title": "Destination",
            "type": "string"
          },
          "lead_ids": {
            "items": {
              "type": "string"
            },
            "title": "Lead Ids",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/ExportStatus",
            "default": "pending"
          },
          "workspace_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace Id"
          }
        },
        "required": [
          "campaign_id",
          "destination",
          "completed_at"
        ],
        "title": "ExportRecordCreate",
        "type": "object"
      },
      "ExportRecordUpdate": {
        "additionalProperties": false,
        "description": "Partial update request.",
        "properties": {
          "campaign_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Campaign Id"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "destination": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Destination"
          },
          "lead_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lead Ids"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExportStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "workspace_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace Id"
          }
        },
        "title": "ExportRecordUpdate",
        "type": "object"
      },
      "ExportStatus": {
        "enum": [
          "pending",
          "processing",
          "completed",
          "failed"
        ],
        "title": "ExportStatus",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RunStatus": {
        "enum": [
          "pending",
          "running",
          "succeeded",
          "failed",
          "cancelled"
        ],
        "title": "RunStatus",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "VA Lead Intelligence API",
    "version": "1.0.0",
    "x-build-sha": "unknown",
    "x-generated-at": "2026-07-15T10:45:31Z"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/exports/": {
      "get": {
        "operationId": "list_export_record_exports__get",
        "parameters": [
          {
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": true,
            "schema": {
              "title": "X-Workspace-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": true,
            "schema": {
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ExportRecord"
                  },
                  "title": "Response List Export Record Exports  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Export Record",
        "tags": [
          "export_record"
        ]
      },
      "post": {
        "operationId": "create_export_record_exports__post",
        "parameters": [
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": true,
            "schema": {
              "title": "X-Workspace-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": true,
            "schema": {
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportRecordCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Export Record",
        "tags": [
          "export_record"
        ]
      }
    },
    "/exports/{export_record_id}": {
      "delete": {
        "operationId": "delete_export_record_exports__export_record_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "export_record_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Export Record Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": true,
            "schema": {
              "title": "X-Workspace-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": true,
            "schema": {
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Export Record",
        "tags": [
          "export_record"
        ]
      },
      "get": {
        "operationId": "get_export_record_exports__export_record_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "export_record_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Export Record Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": true,
            "schema": {
              "title": "X-Workspace-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": true,
            "schema": {
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Export Record",
        "tags": [
          "export_record"
        ]
      },
      "patch": {
        "operationId": "update_export_record_exports__export_record_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "export_record_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Export Record Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": true,
            "schema": {
              "title": "X-Workspace-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": true,
            "schema": {
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportRecordUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Export Record",
        "tags": [
          "export_record"
        ]
      }
    },
...
    "/refresh-requests/": {
      "get": {
        "operationId": "list_discovery_event_refresh_requests__get",
        "parameters": [
          {
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": true,
            "schema": {
              "title": "X-Workspace-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": true,
            "schema": {
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DiscoveryEvent"
                  },
                  "title": "Response List Discovery Event Refresh Requests  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Discovery Event",
        "tags": [
          "discovery_event"
        ]
      },
      "post": {
        "operationId": "create_discovery_event_refresh_requests__post",
        "parameters": [
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": true,
            "schema": {
              "title": "X-Workspace-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": true,
            "schema": {
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiscoveryEventCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscoveryEvent"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Discovery Event",
        "tags": [
          "discovery_event"
        ]
      }
    },
    "/refresh-requests/{discovery_event_id}": {
      "delete": {
        "operationId": "delete_discovery_event_refresh_requests__discovery_event_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "discovery_event_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Discovery Event Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": true,
            "schema": {
              "title": "X-Workspace-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": true,
            "schema": {
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Discovery Event",
        "tags": [
          "discovery_event"
        ]
      },
      "get": {
        "operationId": "get_discovery_event_refresh_requests__discovery_event_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "discovery_event_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Discovery Event Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": true,
            "schema": {
              "title": "X-Workspace-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": true,
            "schema": {
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscoveryEvent"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Discovery Event",
        "tags": [
          "discovery_event"
        ]
      },
      "patch": {
        "operationId": "update_discovery_event_refresh_requests__discovery_event_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "discovery_event_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Discovery Event Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": true,
            "schema": {
              "title": "X-Workspace-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": true,
            "schema": {
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiscoveryEventUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscoveryEvent"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Discovery Event",
        "tags": [
          "discovery_event"
        ]
      }
    },
...
  }
}